Truong601

+++ billiards-manager/routes/web.php
@@ -25,6 +25,7 @@ use App\Http\Controllers\ReportController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\Api\ComboTimeController;
 
 
 Route::get('/', function () {
@@ -56,9 +57,9 @@ Route::prefix('admin')->middleware(['auth', 'admin'])->group(function () {
     Route::get('/combos/{id}/edit', [ComboController::class, 'edit'])->name('admin.combos.edit');
     Route::put('/combos/{id}', [ComboController::class, 'update'])->name('admin.combos.update');
     Route::delete('/combos/{id}', [ComboController::class, 'destroy'])->name('admin.combos.destroy');
-    Route::get('/combos/trash', [ComboController::class, 'trash'])->name('trash');
-    Route::post('/combos/restore/{id}', [ComboController::class, 'restore'])->name('restore');
-    Route::delete('/combos/force-delete/{id}', [ComboController::class, 'forceDelete'])->name('forceDelete');
+    Route::get('/combos/trash', [ComboController::class, 'trash'])->name('admin.combos.trash');
+    Route::post('/combos/restore/{id}', [ComboController::class, 'restore'])->name('admin.combos.restore');
+    Route::delete('/combos/force-delete/{id}', [ComboController::class, 'forceDelete'])->name('admin.combos.forceDelete');
 
 
     // Roles
